requirements.txt
# host environment dependencies

# rest api + sse streaming layer
fastapi
uvicorn[standard]

# model context protocol sdk
mcp

# local ai model client
ollama

# html parsing for ioc extraction
beautifulsoup4

# threat feed retrieval
requests
python-whois

# database
prisma

# env file loading
python-dotenv

# campaign clustering (screenshot perceptual hashing)
imagehash
Pillow
yara-python
